packages/fapilog-tamper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fapilog-tamper"
version = "0.1.0"
description = "Tamper-evident logging add-on for fapilog"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Fapilog", email = "dev@fapilog.dev"}]
dependencies = [
  "fapilog>=0.3.0",
  "cryptography>=41.0.0",
]

[project.optional-dependencies]
signatures = ["pynacl>=1.5.0"]

[project.entry-points."fapilog.integrity"]
tamper-sealed = "fapilog_tamper:TamperSealedPlugin"

[project.scripts]
fapilog-tamper = "fapilog_tamper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fapilog_tamper"]
